centralize choice enums
