+++ app/people/models/role_assignment.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from django.contrib.auth import get_user_model
 from django.db import models
 
-from app.shared.constants.perms import UserRole
+from app.people.choices import UserRole
 
 User = get_user_model()
