+++ app/registry/models/document.py
@@ -8,7 +8,7 @@ from django.contrib.contenttypes.fields import GenericForeignKey
 from django.contrib.contenttypes.models import ContentType
 from django.db import models
 
-from app.shared.constants import DocumentType, StatusDocument
+from app.registry.choices import DocumentType, StatusDocument
 from app.shared.mixins import StatusableMixin, StatusHistory
