+++ app/shared/constants/perms.py
@@ -7,25 +7,9 @@ help the permissions middleware resolve context information."""
 
 from django.db import models
 
+from app.people.choices import UserRole
+
 TEST_PW = "test"
-
-
-class UserRole(models.TextChoices):
-    STUDENT = "student", "Student"
-    PROSPECTIVE_STUDENT = "prospective_student", "Prospective Student"
-    TECHNICIAN = "technician", "Technician"
-    LAB_TECHNICIAN = "lab_technician", "Lab Technician"
-    REGISTRAR = "registrar", "Registrar"
-    ENROLLMENT_OFFICER = "enrollment_officer", "Enrollment Officer"
-    DEAN = "dean", "Dean"
-    CHAIR = "chair", "Chair"
-    LECTURER = "lecturer", "Lecturer"
-    ASSISTANT_PROFESSOR = "assistant_professor", "Assistant Professor"
-    ASSOCIATE_PROFESSOR = "associate_professor", "Associate Professor"
-    PROFESSOR = "professor", "Professor"
-    FACULTY = "faculty", "Faculty"
-    VPAA = "vpaa", "Vpaa"
-    FINANCIALOFFICER = "financial_officer", "Financial Officer"
 
 
 DEFAULT_ROLE_TO_COLLEGE = {
