+++ app/shared/constants/timetable.py
@@ -3,11 +3,4 @@
 Currently this includes the :class:`StatusReservation` enumeration used to
 track the life cycle of a student's reservation of a section."""
 
-from django.db import models
-
-
-class StatusReservation(models.TextChoices):
-    CANCELLED = "cancelled", "Cancelled"
-    PAID = "paid", "Paid"
-    REQUESTED = "requested", "Requested"
-    VALIDATED = "validated", "Validated"
+from app.timetable.choices import StatusReservation
