+++ app/shared/management/populate_helpers/auth.py
@@ -10,7 +10,7 @@ from django.utils import timezone
 from app.academics.models import College
 from app.people.models import RoleAssignment
 from app.shared.constants import DEFAULT_ROLE_TO_COLLEGE, TEST_PW
-from app.shared.constants.perms import UserRole
+from app.people.choices import UserRole
 
 from .utils import log
