+++ app/timetable/choices.py
@@ -1,6 +1,11 @@
-"""Enums module."""
-
+from django.db import models
 from django.db.models import IntegerChoices
+
+class StatusReservation(models.TextChoices):
+    CANCELLED = "cancelled", "Cancelled"
+    PAID = "paid", "Paid"
+    REQUESTED = "requested", "Requested"
+    VALIDATED = "validated", "Validated"
 
 
 class SEMESTER_NUMBER(IntegerChoices):
@@ -13,24 +18,6 @@ class SEMESTER_NUMBER(IntegerChoices):
 class TERM_NUMBER(IntegerChoices):
     FIRST = 1, "First"
     SECOND = 2, "Second"
-
-
-class CREDIT_NUMBER(IntegerChoices):
-    ZERO = 0, "0"
-    ONE = 1, "1"
-    TWO = 2, "2"
-    THREE = 3, "3"
-    FOUR = 4, "4"
-    SIX = 6, "6"
-    TEN = 10, "10"
-
-
-class LEVEL_NUMBER(IntegerChoices):
-    ONE = 1, "freshman"
-    TWO = 2, "sophomore"
-    THREE = 3, "junior"
-    FOUR = 4, "senior"
-    FIVE = 5, "senior"
 
 
 class WEEKDAYS_NUMBER(IntegerChoices):
